fix: code review comments 4

+++ src/schemas/account.ts
@@ -11,8 +11,8 @@ export const sendInsufficientTokensEmailSchema = z.object({
   account_id: accountIdSchema,
   email: z.string().email(),
   first_name: z.string(),
-  available_tokens: z.number().int(), // Allow negative for overdraft
-  required_tokens: z.number().int().positive() // Operations can't require negative tokens
+  available_tokens: z.number(), // Allow negative for overdraft
+  required_tokens: z.number().positive() // Operations can't require negative tokens
 })
 
 // snake_case because the request is coming from the backend server
@@ -20,5 +20,5 @@ export const sendPaymentActivationEmailSchema = z.object({
   account_id: accountIdSchema,
   email: z.string().email(),
   first_name: z.string(),
-  token_balance: z.number().int() // Allow negative for overdraft
+  token_balance: z.number() // Allow negative for overdraft
 })
